+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - More descriptive error messages when a column or association is not found.
 
+### Fixed
+- `Arel::Nodes::Grouping` does not include `Arel::Math`, so operations like `(id + 5) + 3` would fail unexpectedly.
+
 ## [0.2.1] - 2015-03-27
 ### Added
 - Support for subqueries.